Fixes #163, address will be fixed if needed.
The system will try to determine if blog address for enrolled or
changed participant should be changed and apply the change as needed.
A message will be displayed if blog URL needed fixes to be applied.
Moved blog generator determination logic into a standalone function.
Added some more helper functions for the newly added functionality.