References #162, EduFeedr part should be almost done.
NB! Please note that changes have been made to the database tables,
consult docs/UPGRADE.txt for specific details. Added blog_base
to participant, this should allow the system to work with feeds
for categories / labels depending on the generator. In case of
WordPress both permalink and non-permalik approach suffice. Also
changed the way posts and comments feeds for WordPress are
discovered, using the new non-permalink approach as default, as
it seems to be working fine for both permalinks enabled and
disabled. Participant profile URL is hidden from the GUI, but
the page itself is present and partially working. This still needs
testing, only tested with WordPress permalinks enabled category.